+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 
 # [1.4.6] - Unreleased
-- Add Flex::fixed (as is the current api in FLTK's Fl_Flex)
+- Add Flex::fixed (as is the current api in FLTK's Fl_Flex).
+- Add app::fl_msg() (a way to get the fl_msg (the win32 MSG object)).
 
 # [1.4.5] - 2023-06-08
 - Return bool from handle_event.